+++ project/backend/src/main/java/com/example/sbd_zoo/service/EmployeeService.java
@@ -8,6 +8,7 @@ import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.Objects;
 
@@ -79,7 +80,7 @@ public class EmployeeService {
                     employeeRun.setEmployee(employee.getPesel());
                     employeeRunService.addEmployeeRun(employeeRun);
                 }
-                deleteEmployee(id);
+                employeeRepository.deleteById(id);
             }
         } else {
 
@@ -92,7 +93,12 @@ public class EmployeeService {
     }
 
     @Transactional
-    public void deleteEmployee(String id) {
-        employeeRepository.deleteById(id);
+    public void deleteEmployee(String id) throws SQLIntegrityConstraintViolationException {
+        List<Animal> animals = animalService.animalRepository.findAnimalByEmployee(id);
+        if (animals.isEmpty()) {
+            employeeRepository.deleteById(id);
+        } else {
+            throw new SQLIntegrityConstraintViolationException("Nie można usunąć pracownika, ponieważ jest on opiekunem zwierząt.");
+        }
     }
 }
